fix: register with rxkotlin and mvp
