+++ app/src/main/AndroidManifest.xml
@@ -51,13 +51,14 @@
         </activity>
         <activity android:name=".initrate.InitialRatingActivity">
         </activity>
-        <activity android:name=".activity.RegisterActivity" />
-        <activity android:name=".login.LoginActivity" >
+        <activity android:name=".register.RegisterActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".login.LoginActivity" >
         </activity>
         <activity android:name=".activity.SplashActivity" />
         <activity android:name=".activity.MainActivity" />
