+++ app/src/main/java/com/example/cocoman/di/PresenterModule.kt
@@ -4,6 +4,8 @@ import com.example.cocoman.initrate.InitialRatingContract
 import com.example.cocoman.initrate.InitialRatingPresenter
 import com.example.cocoman.login.LoginContract
 import com.example.cocoman.login.LoginPresenter
+import com.example.cocoman.register.RegisterContract
+import com.example.cocoman.register.RegisterPresenter
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
@@ -18,4 +20,7 @@ abstract class PresenterModule {
 
     @Binds
     abstract fun bindInitialRatingPresenter(impl: InitialRatingPresenter): InitialRatingContract.Presenter
+
+    @Binds
+    abstract fun bindRegisterPresenter(impl: RegisterPresenter): RegisterContract.Presenter
 }
